requirements.txt
-i https://pypi.org/simple
anyio==3.6.2 ; python_full_version >= '3.6.2'
beautifulsoup4==4.12.2
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
fastapi==0.95.2
h11==0.14.0 ; python_version >= '3.7'
httptools==0.5.0
idna==3.4 ; python_version >= '3.5'
pydantic==1.10.7 ; python_version >= '3.7'
python-dotenv==1.0.0
pyyaml==6.0
requests==2.30.0
sniffio==1.3.0 ; python_version >= '3.7'
soupsieve==2.4.1 ; python_version >= '3.7'
starlette==0.27.0 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
urllib3==2.0.2 ; python_version >= '3.7'
uvicorn[standard]==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
